Fixed settings dialog being visible on startup on upgraded installations
Versions prior to 2.0 had the index of the error window set to 1, which is used for the settings
window now. Dialogs don't get unregistered so anybody with an old version will still trigger
the function on index 1 if the window was left open before closing. A nice looking separator
on index 1 fixes this without any nasty code. Yay!